requirements.txt
scikit-learn==1.2.1
mysql-connector-python==8.0.12
tqdm==4.64.0
opencv-python-headless
torchvision==0.9.0
click==7.1.2
streamlit==1.12.0
pillow==8.3.2
numpy==1.23.1
pandas==1.4.4
altair==4.1.0
werkzeug==2.2.2
imageio==2.26.0
scikit-image==0.19.3
requests==2.28.1
pydeck==0.7.1
plotly==5.9.0
pyecharts==2.0.3